Use integers instead of longs for min/max values

+++ influxdb-client/src/main/java/net/cpollet/rateme/influxdb/DefaultRate.java
@@ -39,7 +39,7 @@ public class DefaultRate implements Rate {
 
     @Override
     public Number rate() {
-        return Long.parseLong(parse().get("rate"));
+        return Integer.parseInt(parse().get("rate"));
     }
 
     @Override
